+++ waspc/cli/src/Wasp/Cli/Command/Db/Seed.hs
@@ -19,12 +19,12 @@ import Wasp.Cli.Command (Command, CommandError (CommandError), require)
 import Wasp.Cli.Command.Compile (analyze)
 import Wasp.Cli.Command.Message (cliSendMessageC)
 import Wasp.Cli.Command.Require.InWaspProject (InWaspProject (InWaspProject))
-import Wasp.Cli.Util.Apps (getWaspEnvVars)
+import Wasp.Cli.Util.PerService (getWaspEnvVars)
 import Wasp.Cli.Util.PortArgument (defaultAppPorts)
 import Wasp.Generator.DbGenerator.Operations (dbSeed)
 import qualified Wasp.Message as Msg
-import Wasp.Project.Apps (server)
 import Wasp.Project.Common (generatedAppDirInWaspProjectDir)
+import Wasp.Project.PerService (server)
 
 seed :: Maybe String -> Command ()
 seed maybeUserProvidedSeedName = do