+++ waspc/cli/src/Wasp/Cli/Util/PerService.hs
@@ -1,22 +1,22 @@
-module Wasp.Cli.Util.Apps where
+module Wasp.Cli.Util.PerService where
 
 import Network.Socket (PortNumber)
 import Wasp.AppSpec (AppSpec)
 import Wasp.Env (EnvVar)
 import qualified Wasp.Generator.ServerGenerator.Common as Server
 import qualified Wasp.Generator.WebAppGenerator.Common as WebApp
-import Wasp.Project.Apps (Apps (..))
+import Wasp.Project.PerService (PerService (..))
 
-getDevUrlMakers :: AppSpec -> Apps (PortNumber -> String)
+getDevUrlMakers :: AppSpec -> PerService (PortNumber -> String)
 getDevUrlMakers spec =
-  Apps
+  PerService
     { client = WebApp.getDevClientUrl spec,
       server = Server.makeDevServerUrl
     }
 
-getWaspEnvVars :: AppSpec -> Apps PortNumber -> Apps [EnvVar]
+getWaspEnvVars :: AppSpec -> PerService PortNumber -> PerService [EnvVar]
 getWaspEnvVars spec ports =
-  Apps
+  PerService
     { client = WebApp.getDevClientEnvVars,
       server = Server.getDevServerEnvVars
     }